refactor common responses and osdi helpers
