+++ lib/osdi/index.js
@@ -1,0 +1,4 @@
+module.exports = {
+  request: require('./request-helper'),
+  response: require('./response-helper')
+};
