+++ lib/osdi/response-helper.js
@@ -1,4 +1,4 @@
-var config = require('../config');
+var config = require('../../config');
 
 var root = config.get('apiEndpoint');
 
@@ -77,6 +77,40 @@ function getPaginationOptions(req) {
   return options;
 }
 
+function unauthorized(res) {
+  return function() {
+    return res.status(401).end();
+  };
+};
+
+function badRequest(res, type) {
+  return function(error) {
+    var response_code = 500;
+    if (!error) {
+      response_code = 400;
+    }
+
+    var answer = {
+      'request_type': 'atomic',
+      'response_code': response_code,
+      'resource_status': [
+        {
+          'resource': 'osd:' + type,
+          'response_code': response_code,
+          'errors': [
+            {
+              'code': 'UNKNOWN',
+              'description': 'Translating VAN errors is not yet supported.'
+            }
+          ]
+        }
+      ]
+    };
+
+    return res.status(response_code).send(answer);
+  };
+}
+
 module.exports = {
   createCommonItem: createCommonItem,
   createPaginatedItem: createPaginatedItem,
@@ -84,5 +118,7 @@ module.exports = {
   addLink: addLink,
   addSelfLink: addSelfLink,
   addIdentifier: addIdentifier,
-  getPaginationOptions: getPaginationOptions
+  getPaginationOptions: getPaginationOptions,
+  unauthorized: unauthorized,
+  badRequest: badRequest
 };
