+++ routes/questions.js
@@ -1,55 +1,20 @@
 
 var contentType = require('../middleware/contentType'),
     config = require('../config'),
-    auth = require('basic-auth'),
     ngpvanAPIClient = require('../lib/ngpvanapi-client'),
-    osdi = require('../lib/osdi-response-helper');
+    osdi = require('../lib/osdi');
 
 var vanEndpoint = config.get('vanEndpoint');
 
-var unauthorized = function(res) {
-  return function() {
-    return res.status(401).end();
-  };
-};
-
-function badRequest(res) {
-  return function(error) {
-    var response_code = 500;
-    if (!error) {
-      response_code = 400;
-    }
-
-    var answer = {
-      'request_type': 'atomic',
-      'response_code': response_code,
-      'resource_status': [
-        {
-          'resource': 'osd:questions',
-          'response_code': response_code,
-          'errors': [
-            {
-              'code': 'UNKNOWN',
-              'description': 'Translating VAN errors is not yet supported.'
-            }
-          ]
-        }
-      ]
-    };
-
-    return res.status(response_code).send(answer);
-  };
-}
-
 function translateSurveyQuestionToQuestion(surveyQuestion) {
-    var answer = osdi.createCommonItem(
+    var answer = osdi.response.createCommonItem(
       surveyQuestion.mediumName,
       surveyQuestion.scriptQuestion);
     answer.title = surveyQuestion.name;
     answer.summary = answer.description;
     answer.question_type = 'SingleChoice';
-    osdi.addIdentifier(answer, 'VAN:' + surveyQuestion.surveyQuestionId);
-    osdi.addSelfLink(answer, 'questions', surveyQuestion.surveyQuestionId);
+    osdi.response.addIdentifier(answer, 'VAN:' + surveyQuestion.surveyQuestionId);
+    osdi.response.addSelfLink(answer, 'questions', surveyQuestion.surveyQuestionId);
 
     answer.responses = (surveyQuestion.responses || []).map(function(response) {
       return {
@@ -81,15 +46,15 @@ function getOne(req, res) {
     return res.status(200).send(answer);
   };
 
-  var credentials = getCredentials(req);
+  var credentials = osdi.request.getCredentials(req);
 
   ngpvanAPIClient.surveyQuestions.getOne(vanEndpoint,
     credentials.apiKey, credentials.dbMode, id,
-    unauthorized(res), badRequest(res), success);
+    osdi.response.unauthorized(res), osdi.response.badRequest(res), success);
 }
 
 function getAll(req, res) {
-  var pagination = osdi.getPaginationOptions(req);
+  var pagination = osdi.response.getPaginationOptions(req);
 
   var success = function(surveyQuestions) {
     if (!(surveyQuestions && surveyQuestions.items)) {
@@ -101,32 +66,20 @@ function getAll(req, res) {
     var totalRecords = surveyQuestions.count;
 
     var totalPages = Math.ceil(totalRecords / perPage);
-    var answer = osdi.createPaginatedItem(page, perPage, totalPages,
+    var answer = osdi.response.createPaginatedItem(page, perPage, totalPages,
       totalRecords, 'questions');
 
     var items = surveyQuestions.items;
 
-    osdi.addEmbeddedItems(answer, items, translateSurveyQuestionToQuestion);
+    osdi.response.addEmbeddedItems(answer, items, translateSurveyQuestionToQuestion);
     return res.status(200).send(answer);
   };
 
-  var credentials = getCredentials(req);
+  var credentials = osdi.request.getCredentials(req);
 
   ngpvanAPIClient.surveyQuestions.getAll(vanEndpoint, pagination,
     credentials.apiKey, credentials.dbMode,
-    unauthorized(res), badRequest(res), success);
-}
-
-function getCredentials(req) {
-  var user = auth(req);
-  var pass = user.pass;
-
-  if (typeof pass !== 'string') {
-    return {};
-  }
-
-  var parts = pass.split('|');
-  return { apiKey: parts[0], dbMode: parts[1] };
+    osdi.response.unauthorized(res), osdi.response.badRequest(res), success);
 }
 
 module.exports = function (app) {
